app/src/androidTest/java/com/antoniofalcescu/licenta/LocalDatabaseTest.kt
package com.antoniofalcescu.licenta

import android.content.Context
import androidx.room.Room
import androidx.test.core.app.ApplicationProvider
import androidx.test.ext.junit.runners.AndroidJUnit4
import com.antoniofalcescu.licenta.repository.roomDatabase.accessToken.AccessToken
import com.antoniofalcescu.licenta.repository.roomDatabase.accessToken.AccessTokenDao
import com.antoniofalcescu.licenta.repository.roomDatabase.LocalDatabase
import org.junit.After
import org.junit.Assert.assertEquals
import org.junit.Before
import org.junit.Test
import org.junit.runner.RunWith
import java.io.IOException

@RunWith(AndroidJUnit4::class)
class LocalDatabaseTest {

    private lateinit var accessTokenDao: AccessTokenDao
    private lateinit var db: LocalDatabase

    @Before
    fun createDb() {
        val context = ApplicationProvider.getApplicationContext<Context>()
        db = Room.inMemoryDatabaseBuilder(context, LocalDatabase::class.java).allowMainThreadQueries().build()

        accessTokenDao = db.accessTokenDao
    }

    @After
    @Throws(IOException::class)
    fun closeDb() {
        db.close()
    }

    @Test
    @Throws(Exception::class)
    fun saveAndGetAccessToken() {
        val savedAccessToken = AccessToken(value = "test-token")
        accessTokenDao.save(savedAccessToken)
        val gottenAccessToken = accessTokenDao.get()
        assertEquals(gottenAccessToken.value, savedAccessToken.value)
    }
}